% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikr-package.R
\docType{data}
\name{bikesharedata}
\alias{bikesharedata}
\title{bikesharedata: Pittsburgh Bike Share trips for Q1 of 2021}
\format{
A csv document with 14,619 cases of 10 variables:
\describe{
\item{trip.id}{ID Number for the individual trip}
\item{Starttime}{The date/time the bikeshare trip began}
\item{Stoptime}{The date/time the bikeshare trip ended}
\item{Bikeid}{The BikeShare bike's identifier number}
\item{Tripduration}{The length of the trip in minutes}
\item{From.station.id}{The origin hub ID number}
\item{From.station.name}{The origin hub name/location}
\item{To.station.id}{The destination hub ID number}
\item{To.station.name}{The destination hub name/location}
\item{Usertype}{Whether the rider is a customer or subscriber}
}
}
\source{
Adam S Carroll & Pittsburgh Healthy Ride BikeShare, (https://www.kaggle.com/adamscarroll/pittsburgh-healthy-ride-bikeshare-rentals-2021-q1/metadata)
}
\description{
A dataset containing records of all Pittsburgh Bike Share trips for Q1 of 2021
including start/end times, durations, origins and destinations, and user types.
}
\keyword{bike}
\keyword{datasets,}
\keyword{share}
