% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikr-package.R
\docType{package}
\name{bikr-package}
\alias{bikr}
\alias{bikr-package}
\title{bikr: About the Package (One Line, Title Case)}
\description{
This package provides an example BikeShare trip dataset from Pittsburgh BikeShare (available for download at []()), as well as several scripts to analyze these trips. Package scripts can produce maps and statistical analysis (graphs, histograms) from trip data.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{http://github.com/vanveghd/bikr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
