% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikr-package.R
\docType{package}
\name{bikrr-package}
\alias{bikrr-package}
\title{bikr: A package for basic BikeShare trip analysis.}
\usage{
bikr is free for use and manipulation, and can be downloaded from the Github Repository. However, the dataset in the package is not the property of the author, and proper attribution of the data is necessary.
}
\description{
This package provides a range of functions varying in depth of analysis,
for a sample BikeShare Dataset available publicly
from (https://www.kaggle.com/adamscarroll/pittsburgh-healthy-ride-bikeshare-rentals-2021-q1/metadata).
The package also provides several tools and scripts for visualizing route distribution,
characteristics, lengths, as well as cross-analyzing all of these characteristics by ride-types.
}
\references{
\url{https://github.com/vanveghd/bikr}
}
\author{
Daniel Van Veghel, School of Earth, Environment and Society, McMaster University \email{vanveghd@mcmaster.ca}
}
\keyword{Analysis,}
\keyword{Behaviour}
\keyword{BikeShare,}
\keyword{Travel}
\keyword{statistics,}
