% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means.R
\name{compare_means}
\alias{compare_means}
\title{Determine the statistical significance of the difference in two different sample means.}
\usage{
compare_means(mean1, mean2, var1, var2, n1, n2)
}
\arguments{
\item{mean1}{The first sample mean}

\item{mean2}{The second sample mean}

\item{var1}{The first sample variance}

\item{var2}{the second sample variance}

\item{n1}{The first sample size}

\item{n2}{The second sample size}
}
\value{
The t value of the associated small two-sample t-test, and the associated p-value for this t-value.
}
\description{
Determine the statistical significance of the difference in two different sample means.
}
