% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_route_durations.R
\name{visualize_route_durations}
\alias{visualize_route_durations}
\title{Visualize the Trip Durations field of the BikeShare Data}
\usage{
visualize_route_durations()
}
\description{
Visualize the Trip Durations field of the BikeShare Data
}
